changed namespace for IEX and moved files around
