+++ services/api-service/lib/iex/index.ts
@@ -1,0 +1,6 @@
+export * as CryptoSymbols from './crypto-symbols.service';
+export * as HistoricBatch from './historic-batch-chart.service';
+export * as HistoricPrices from './historic-prices.service';
+export * as IEXCloud from './iexcloud.service';
+export * as Quote from './quote.service';
+export * as Symbols from './symbol.service';
