+++ services/api-service/test/iex/crypto-symbols.service.spec.ts
@@ -1,13 +1,13 @@
 import { describe } from 'mocha';
 import { assert, expect } from 'chai';
-import { tokenPlugin } from '../lib/iexcloud.service';
-import * as HistoricPrices from '../lib/historic-prices.service';
+import { tokenPlugin } from '../../lib/iex/iexcloud.service';
+import * as CryptoSymbols from '../../lib/iex/crypto-symbols.service';
 import Sinon from 'sinon';
-import { HappyPathMock } from '../mocks/historic-prices.mock';
+import { HappyPathMock } from '../../mocks/crypto-symbols.mock';
 
-const STOCK_SYMBOL = 'AAPL';
+const STOCK_SYMBOL = 'BTCUSD';
 
-describe('HistoricPrices::', function () {
+describe('CryptoSymbols::', function () {
     let sandbox: Sinon.SinonSandbox;
     let stubs: any = {};
 
@@ -23,24 +23,18 @@ describe('HistoricPrices::', function () {
     });
 
     it('Should run happy path', async () => {
-        const resp: any = await HistoricPrices.historic({
-            symbol: STOCK_SYMBOL,
-            timeframe: '1d'
-        });
+        const resp: any = await CryptoSymbols.cryptoSymbols();
         assert(resp);
         expect(resp).to.be.a('array');
         const quote = resp[0];
         expect(quote.symbol).to.be.a('string').and.equal(STOCK_SYMBOL);
-        expect(quote.high).to.be.a('number').and.equal(0);
+        expect(quote.name).to.be.a('string').and.equal("Bitcoin to USD");
     });
 
     it('Should run unhappy path', async () => {
         stubs.gotGetStub.rejects(Error('err'));
         try {
-            const resp: any = await HistoricPrices.historic({
-                symbol: STOCK_SYMBOL,
-                timeframe: '1d'
-            });
+            const resp: any = await CryptoSymbols.cryptoSymbols();
             assert(!resp);
         } catch (err) {
             assert(err);
