+++ services/api-service/test/iex/historic-prices.service.spec.ts
@@ -1,13 +1,13 @@
 import { describe } from 'mocha';
 import { assert, expect } from 'chai';
-import { tokenPlugin } from '../lib/iexcloud.service';
-import * as HistoricBatchPrices from '../lib/historic-batch-chart.service';
+import { tokenPlugin } from '../../lib/iex/iexcloud.service';
+import * as HistoricPrices from '../../lib/iex/historic-prices.service';
 import Sinon from 'sinon';
-import { HappyPathMock } from '../mocks/historical-batch.mock';
+import { HappyPathMock } from '../../mocks/historic-prices.mock';
 
 const STOCK_SYMBOL = 'AAPL';
 
-describe('HistoricalBatchChart::', function () {
+describe('HistoricPrices::', function () {
     let sandbox: Sinon.SinonSandbox;
     let stubs: any = {};
 
@@ -23,16 +23,13 @@ describe('HistoricalBatchChart::', function () {
     });
 
     it('Should run happy path', async () => {
-        const resp: any = await HistoricBatchPrices.historicBatch({
-            symbols: [STOCK_SYMBOL],
+        const resp: any = await HistoricPrices.historic({
+            symbol: STOCK_SYMBOL,
             timeframe: '1d'
         });
         assert(resp);
-        const stockBatchData = resp[STOCK_SYMBOL];
-        expect(stockBatchData).to.be.a('object');
-        const stockBatchChartData = stockBatchData.chart;
-        expect(stockBatchChartData).to.be.a('array');
-        const quote = stockBatchChartData[0];
+        expect(resp).to.be.a('array');
+        const quote = resp[0];
         expect(quote.symbol).to.be.a('string').and.equal(STOCK_SYMBOL);
         expect(quote.high).to.be.a('number').and.equal(0);
     });
@@ -40,8 +37,8 @@ describe('HistoricalBatchChart::', function () {
     it('Should run unhappy path', async () => {
         stubs.gotGetStub.rejects(Error('err'));
         try {
-            const resp: any = await HistoricBatchPrices.historicBatch({
-                symbols: [STOCK_SYMBOL],
+            const resp: any = await HistoricPrices.historic({
+                symbol: STOCK_SYMBOL,
                 timeframe: '1d'
             });
             assert(!resp);
